size_t changed into int at line 391

+++ FVMtools/extract.cc
@@ -388,7 +388,7 @@ volumePart::volumePart(string out_dir, string iteration, string casename,
       int cnt1 = 0 ;
       int cnt2 = 0 ;
       int cnt = 0 ;
-      for(size_t i=0;i<ngenc;++i) {
+      for(int i=0;i<ngenc;++i) {
         bool blank = true ;
         int nf = GeneralCellNfaces[i] ;
         for(int f=0;f<nf;++f) {
